Add AddDocument API to DocumentDock and update Notepad sample
