--- samples/Notepad/ViewModels/Docks/FilesDocumentDock.cs
@@ -1,34 +1,0 @@
-﻿using System.Text;
-using CommunityToolkit.Mvvm.Input;
-using Dock.Model.Mvvm.Controls;
-using Notepad.ViewModels.Documents;
-
-namespace Notepad.ViewModels.Docks;
-
-public class FilesDocumentDock : DocumentDock
-{
-    public FilesDocumentDock()
-    {
-        CreateDocument = new RelayCommand(CreateNewDocument);
-    }
-
-    private void CreateNewDocument()
-    {
-        if (!CanCreateDocument)
-        {
-            return;
-        }
-
-        var document = new FileViewModel()
-        {
-            Path = string.Empty,
-            Title = "Untitled",
-            Text = "",
-            Encoding = Encoding.Default.WebName
-        };
-
-        Factory?.AddDockable(this, document);
-        Factory?.SetActiveDockable(document);
-        Factory?.SetFocusedDockable(this, document);
-    }
-}
